- Fixed: R_GetOneSkyColumn() and R_GetTwoSkyColumns are mulscaling an   unsigned integer that can use all 32 bits. They must therefore use   the unsigned mul instruction rather than the signed imul instruction. - Fixed several signed/unsigned comparison and possibly uninitialized   variable warnings flagged by GCC.
SVN r1965 (trunk)

+++ src/p_buildmap.cpp
@@ -624,11 +624,11 @@ static void LoadWalls (walltype *walls, int numwalls, sectortype *bsec)
 		int sidenum = int(intptr_t(lines[linenum].sidedef[1]));
 		if (bsec->floorstat & 64)
 		{ // floor is aligned to first wall
-			R_AlignFlat (linenum, sidenum == (DWORD)bsec->wallptr, 0);
+			R_AlignFlat (linenum, sidenum == bsec->wallptr, 0);
 		}
 		if (bsec->ceilingstat & 64)
 		{ // ceiling is aligned to first wall
-			R_AlignFlat (linenum, sidenum == (DWORD)bsec->wallptr, 0);
+			R_AlignFlat (linenum, sidenum == bsec->wallptr, 0);
 		}
 	}
 	for(i = 0; i < numsides; i++)
